Bugfix & Compile

+++ Core/Inc/printhead.h
@@ -20,47 +20,41 @@
 #include <limits>
 #include <algorithm>
 
-class Printhead
-{
+class Printhead {
 public:
-	Printhead(TIM_HandleTypeDef *htim, uint32_t timChannel) : htim(htim), channel(timChannel)
-	{
+	Printhead(TIM_HandleTypeDef *htim, uint32_t timChannel) :
+			htim(htim), channel(timChannel) {
 	}
 
-	struct param
-	{
+	struct param {
 		uint16_t period = PRINTHEAD_PERIOD;
 		uint8_t dutyCycle = PRINTHEAD_DUTY_CYCLE;
 	} param;
 
 	void setParam(uint16_t newPeriod = PRINTHEAD_PERIOD, uint8_t newDutyCycle =
-															 PRINTHEAD_DUTY_CYCLE)
-	{
+	PRINTHEAD_DUTY_CYCLE) {
 		param.period = newPeriod;
-		param.dutyCycle = std::clamp(newDutyCycle, 0, 100);
-		htim->Instance->ARR = std::clamp((uint16_t)(param.period * 10 - 1),
-										 std::numeric_limits<uint16_t>::min(),
-										 std::numeric_limits<uint16_t>::max());
-		htim->Instance->CCR1 = std::clamp((uint16_t)(param.dutyCycle * param.period * 10 / 100 - 1),
-										  std::numeric_limits<uint16_t>::min(),
-										  std::numeric_limits<uint16_t>::max());
+		param.dutyCycle = std::clamp(newDutyCycle, (uint8_t) 0, (uint8_t) 100);
+		htim->Instance->ARR = std::clamp((uint16_t) (param.period * 10 - 1),
+				std::numeric_limits<uint16_t>::min(),
+				std::numeric_limits<uint16_t>::max());
+		htim->Instance->CCR1 = std::clamp(
+				(uint16_t) (param.dutyCycle * param.period * 10 / 100 - 1),
+				std::numeric_limits<uint16_t>::min(),
+				std::numeric_limits<uint16_t>::max());
 	}
-	void start()
-	{
+	void start() {
 		HAL_TIM_PWM_Start(htim, channel);
 		active = true;
 	}
-	void stop()
-	{
+	void stop() {
 		HAL_TIM_PWM_Stop(htim, channel);
 		active = false;
 	}
-	bool isActive()
-	{
+	bool isActive() {
 		return active;
 	}
-	void init()
-	{
+	void init() {
 		stop();
 		setParam();
 	}
